Add basic PolyGerrit UI
Due to lack of an extension point in the related changes list, we add
the CRDs underneath votes. This is also where it was for the GWT UI.

This basic variant of the UI shows the CRDs and provides links. But it
does not yet color them red if cycles are detected.

Change-Id: I019d3b35de2933abecb5caef0613a02904c8eb68